python/examples/integrations/flask_streaming/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
whylogs[whylabs]>=1
python-dotenv==0.19.0
pandas==1.3.2
joblib==1.0.1
flask_cors==3.0.10
flasgger==0.9.5
pydantic==1.8.2
Flask-Pydantic==0.8.0
utils==1.0.1
# APScheduler==3.7.0
scikit-learn==1.1.1
requests==2.22.0
